Move request method to passport utils
