+++ server/routes/auth/providers/azure.js
@@ -1,14 +1,13 @@
 // @flow
 import passport from "@outlinewiki/koa-passport";
 import { Strategy as AzureStrategy } from "@outlinewiki/passport-azure-ad-oauth2";
-import fetch from "fetch-with-proxy";
 import jwt from "jsonwebtoken";
 import Router from "koa-router";
 import accountProvisioner from "../../../commands/accountProvisioner";
 import env from "../../../env";
 import { MicrosoftGraphError } from "../../../errors";
 import passportMiddleware from "../../../middlewares/passport";
-import { StateStore } from "../../../utils/passport";
+import { StateStore, request } from "../../../utils/passport";
 
 const router = new Router();
 const providerName = "azure";
@@ -17,17 +16,6 @@ const AZURE_CLIENT_SECRET = process.env.AZURE_CLIENT_SECRET;
 const AZURE_RESOURCE_APP_ID = process.env.AZURE_RESOURCE_APP_ID;
 
 const scopes = [];
-
-export async function request(endpoint: string, accessToken: string) {
-  const response = await fetch(endpoint, {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      "Content-Type": "application/json",
-    },
-  });
-  return response.json();
-}
 
 export const config = {
   name: "Microsoft",
